Корректировка стадий игры покер.
Пакет combination.determinant:

    - Добавление определителей:
      * Pair
      * High Card
      Данные определители не пересекаются с Dublicates

Пакет core.player:

    - Микроизменения для пакета core.stage

Пакет core.stage:

    - Реализованы Trade и Showdown (коряво).
    - Другие стадии изменены из-за добавления ActionRequirements.

Пакет core:

    - Добавление ActionRequirements определяющий корректность действия
      и частичного его исполнения.

+++ src/main/java/stud/task/core/player/Storage.java
@@ -28,7 +28,7 @@ public class Storage {
         purse += sum;
     }
 
-    public double getSum() {
+    public long getSum() {
         return purse;
     }
